+++ packages/core-js-compat/src/modules-by-versions.mjs
@@ -323,4 +323,7 @@ export default {
     'es.weak-map.get-or-insert',
     'es.weak-map.get-or-insert-computed',
   ],
+  '3.50': [
+    'es.regexp.has-indices',
+  ],
 };